Split multifile generic generated C tests
