+++ tests/integration/generic_specializations/multifile_generated_c.rs
@@ -1,229 +1,6 @@
 use crate::support::*;
 
-#[test]
-fn multi_file_generic_specializations_do_not_emit_unspecialized_c_symbols() {
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("multi_file_generic/main.zen"));
-    assert!(c_source.contains("typedef struct Option_i32 Option_i32;"));
-    assert!(c_source.contains("typedef struct Result_i32_str Result_i32_str;"));
-    assert!(c_source.contains("int32_t unwrap_option_i32(Option_i32 value, int32_t fallback)"));
-    assert!(
-        c_source.contains("int32_t unwrap_result_i32_str(Result_i32_str value, int32_t fallback)")
-    );
-    assert!(c_source.contains("unwrap_option_i32(some, 0LL)"));
-    assert!(c_source.contains("unwrap_result_i32_str(err, 9LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "unwrap_option_i32");
-    assert_c_call_resolves_to_definition(&c_source, "unwrap_result_i32_str");
-    assert!(!c_source.contains("Option_T"));
-    assert!(!c_source.contains("Result_T"));
-    assert!(!c_source.contains("T unwrap_option"));
-    assert!(!c_source.contains("T unwrap_result"));
-    assert!(!c_source.contains("unwrap_option(some"));
-    assert!(!c_source.contains("unwrap_result(err"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_generic_enum_method/main.zen"),
-    );
-    assert!(c_source.contains("typedef struct Option_i32 Option_i32;"));
-    assert!(c_source.contains("int32_t Option_unwrap_or_i32(Option_i32 self, int32_t fallback)"));
-    assert!(c_source.contains("Option_unwrap_or_i32(some, 0LL)"));
-    assert!(c_source.contains("Option_unwrap_or_i32(none, 89LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "Option_unwrap_or_i32");
-    assert!(!c_source.contains("Option_T"));
-    assert!(!c_source.contains("T Option_unwrap_or"));
-    assert!(!c_source.contains("Option_unwrap_or(some"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_generic_result_enum_method/main.zen"),
-    );
-    assert!(c_source.contains("typedef struct Result_i32_str Result_i32_str;"));
-    assert!(c_source
-        .contains("int32_t Result_unwrap_or_i32_str(Result_i32_str self, int32_t fallback)"));
-    assert!(c_source.contains("Result_unwrap_or_i32_str(ok, 0LL)"));
-    assert!(c_source.contains("Result_unwrap_or_i32_str(err, 144LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "Result_unwrap_or_i32_str");
-    assert!(!c_source.contains("Result_T"));
-    assert!(!c_source.contains("T Result_unwrap_or"));
-    assert!(!c_source.contains("Result_unwrap_or(err"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_generic_result_enum_multi_specialization/main.zen"),
-    );
-    assert!(c_source.contains("typedef struct Result_i32_str Result_i32_str;"));
-    assert!(c_source.contains("typedef struct Result_bool_str Result_bool_str;"));
-    assert!(c_source
-        .contains("int32_t Result_unwrap_or_i32_str(Result_i32_str self, int32_t fallback)"));
-    assert!(
-        c_source.contains("bool Result_unwrap_or_bool_str(Result_bool_str self, bool fallback)")
-    );
-    assert!(c_source.contains("Result_unwrap_or_i32_str(ok_int, 0LL)"));
-    assert!(c_source.contains("Result_unwrap_or_i32_str(err_int, 144LL)"));
-    assert!(c_source.contains("Result_unwrap_or_bool_str(ok_bool, true)"));
-    assert!(c_source.contains("Result_unwrap_or_bool_str(err_bool, true)"));
-    assert_c_call_resolves_to_definition(&c_source, "Result_unwrap_or_i32_str");
-    assert_c_call_resolves_to_definition(&c_source, "Result_unwrap_or_bool_str");
-    assert_c_function_definition_count(&c_source, "Result_unwrap_or_i32_str", 1);
-    assert_c_function_definition_count(&c_source, "Result_unwrap_or_bool_str", 1);
-    assert!(!c_source.contains("Result_T"));
-    assert!(!c_source.contains("T Result_unwrap_or"));
-    assert!(!c_source.contains("Result_unwrap_or(err"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_generic_imported_type_dependency/main.zen"),
-    );
-    assert!(c_source.contains("typedef struct Holder_i32 Holder_i32;"));
-    assert!(c_source.contains("int32_t Holder_get_i32(Holder_i32 self)"));
-    assert!(c_source.contains("int32_t get_held_i32(int32_t value)"));
-    assert!(c_source.contains("const Holder_i32 holder = (Holder_i32){ .value = value }"));
-    assert!(c_source.contains("Holder_get_i32(holder)"));
-    assert!(c_source.contains("get_held_i32(73LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "Holder_get_i32");
-    assert_c_call_resolves_to_definition(&c_source, "get_held_i32");
-    assert!(!c_source.contains("Holder_T"));
-    assert!(!c_source.contains("T Holder_get"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_generic_imported_worklist_chain/main.zen"),
-    );
-    assert!(c_source.contains("int32_t inner_i32(int32_t value)"));
-    assert!(c_source.contains("int32_t middle_i32(int32_t value)"));
-    assert!(c_source.contains("int32_t outer_i32(int32_t value)"));
-    assert!(c_source.contains("inner_i32(value)"));
-    assert!(c_source.contains("middle_i32(value)"));
-    assert!(c_source.contains("outer_i32(83LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "inner_i32");
-    assert_c_call_resolves_to_definition(&c_source, "middle_i32");
-    assert_c_call_resolves_to_definition(&c_source, "outer_i32");
-    assert!(!c_source.contains("T inner"));
-    assert!(!c_source.contains("T middle"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_generic_imported_transitive_dependency/main.zen"),
-    );
-    assert!(c_source.contains("int32_t inner_i32(int32_t value)"));
-    assert!(c_source.contains("int32_t middle_i32(int32_t value)"));
-    assert!(c_source.contains("int32_t outer_i32(int32_t value)"));
-    assert!(c_source.contains("inner_i32(value)"));
-    assert!(c_source.contains("middle_i32(value)"));
-    assert!(c_source.contains("outer_i32(89LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "inner_i32");
-    assert_c_call_resolves_to_definition(&c_source, "middle_i32");
-    assert_c_call_resolves_to_definition(&c_source, "outer_i32");
-    assert!(!c_source.contains("T inner"));
-    assert!(!c_source.contains("T middle"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("multi_file_type_impl/main.zen"));
-    assert!(c_source.contains("int32_t Box_get_i32(Box_i32 self)"));
-    assert!(c_source.contains("Box_get_i32(box)"));
-    assert_c_call_resolves_to_definition(&c_source, "Box_get_i32");
-    assert!(!c_source.contains("Box_T"));
-    assert!(!c_source.contains("T Box_get"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_type_impl_imported_type_dependency/main.zen"),
-    );
-    assert!(c_source.contains("typedef struct Holder_i32 Holder_i32;"));
-    assert!(c_source.contains("int32_t Holder_get_i32(Holder_i32 self)"));
-    assert!(c_source.contains("int32_t Box_get_held_i32(Box_i32 self)"));
-    assert!(c_source.contains("const Holder_i32 holder = (Holder_i32){ .value = self.value }"));
-    assert!(c_source.contains("Holder_get_i32(holder)"));
-    assert!(c_source.contains("Box_get_held_i32(box)"));
-    assert_c_call_resolves_to_definition(&c_source, "Holder_get_i32");
-    assert_c_call_resolves_to_definition(&c_source, "Box_get_held_i32");
-    assert!(!c_source.contains("Holder_T"));
-    assert!(!c_source.contains("T Holder_get"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_type_impl_return_enum_dependency/main.zen"),
-    );
-    assert!(c_source.contains("typedef struct Option_i32 Option_i32;"));
-    assert!(c_source.contains("Option_i32 Box_wrap_i32(Box_i32 self)"));
-    assert!(c_source.contains("int32_t Box_value_or_i32(Box_i32 self, int32_t fallback)"));
-    assert!(c_source.contains("Box_wrap_i32(self)"));
-    assert!(c_source.contains("Box_value_or_i32(box, 0LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "Box_wrap_i32");
-    assert_c_call_resolves_to_definition(&c_source, "Box_value_or_i32");
-    assert!(!c_source.contains("Option_T"));
-    assert!(!c_source.contains("T Box_wrap"));
-
-    let c_source =
-        compile_to_c_with_generated_call_check(&test_dir().join("multi_file_type_method/main.zen"));
-    assert!(c_source.contains("int32_t Point_keep_i32(Point self, int32_t value)"));
-    assert!(c_source.contains("Point_keep_i32(point, 13LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "Point_keep_i32");
-    assert!(!c_source.contains("T Point_keep"));
-    assert!(!c_source.contains("Point_keep(point"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_type_method_worklist/main.zen"),
-    );
-    assert!(c_source.contains("int32_t inner_i32(int32_t value)"));
-    assert!(c_source.contains("int32_t Box_get_inner_i32(Box_i32 self)"));
-    assert!(c_source.contains("inner_i32(self.value)"));
-    assert!(c_source.contains("Box_get_inner_i32(box)"));
-    assert_c_call_resolves_to_definition(&c_source, "inner_i32");
-    assert_c_call_resolves_to_definition(&c_source, "Box_get_inner_i32");
-    assert!(!c_source.contains("T inner"));
-    assert!(!c_source.contains("inner_T"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_type_method_method_dependency/main.zen"),
-    );
-    assert!(c_source.contains("int32_t Box_inner_i32(Box_i32 self)"));
-    assert!(c_source.contains("int32_t Box_get_inner_i32(Box_i32 self)"));
-    assert!(c_source.contains("Box_inner_i32(self)"));
-    assert!(c_source.contains("Box_get_inner_i32(box)"));
-    assert_c_call_resolves_to_definition(&c_source, "Box_inner_i32");
-    assert_c_call_resolves_to_definition(&c_source, "Box_get_inner_i32");
-    assert!(!c_source.contains("T Box_inner"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_type_method_imported_dependency/main.zen"),
-    );
-    assert!(c_source.contains("int32_t inner_i32(int32_t value)"));
-    assert!(c_source.contains("int32_t Box_get_inner_i32(Box_i32 self)"));
-    assert!(c_source.contains("inner_i32(self.value)"));
-    assert!(c_source.contains("Box_get_inner_i32(box)"));
-    assert_c_call_resolves_to_definition(&c_source, "inner_i32");
-    assert_c_call_resolves_to_definition(&c_source, "Box_get_inner_i32");
-    assert!(!c_source.contains("T inner"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_type_method_return_enum_dependency/main.zen"),
-    );
-    assert!(c_source.contains("typedef struct Option_i32 Option_i32;"));
-    assert!(c_source.contains("Option_i32 Box_wrap_i32(Box_i32 self)"));
-    assert!(c_source.contains("int32_t Box_value_or_i32(Box_i32 self, int32_t fallback)"));
-    assert!(c_source.contains("Box_wrap_i32(self)"));
-    assert!(c_source.contains("Box_value_or_i32(box, 0LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "Box_wrap_i32");
-    assert_c_call_resolves_to_definition(&c_source, "Box_value_or_i32");
-    assert!(!c_source.contains("Option_T"));
-    assert!(!c_source.contains("T Box_wrap"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_type_method_nested_result_dependency/main.zen"),
-    );
-    assert!(c_source.contains("typedef struct Option_i32 Option_i32;"));
-    assert!(c_source.contains("typedef struct Result_Option_i32_str Result_Option_i32_str;"));
-    assert!(c_source.contains("Result_Option_i32_str Box_wrap_result_i32(Box_i32 self)"));
-    assert!(c_source.contains("Box_wrap_result_i32(box)"));
-    assert_c_call_resolves_to_definition(&c_source, "Box_wrap_result_i32");
-    assert!(!c_source.contains("Option_T"));
-    assert!(!c_source.contains("Result_Option_T"));
-    assert!(!c_source.contains("T Box_wrap_result"));
-
-    let c_source = compile_to_c_with_generated_call_check(
-        &test_dir().join("multi_file_imported_generic_function_return_enum_dependency/main.zen"),
-    );
-    assert!(c_source.contains("typedef struct Option_i32"));
-    assert!(c_source.contains("Option_i32 wrap_i32(int32_t value)"));
-    assert!(c_source.contains("int32_t unwrap_i32(Option_i32 value, int32_t fallback)"));
-    assert!(c_source.contains("wrap_i32(107LL)"));
-    assert!(c_source.contains("unwrap_i32(value, 0LL)"));
-    assert_c_call_resolves_to_definition(&c_source, "wrap_i32");
-    assert_c_call_resolves_to_definition(&c_source, "unwrap_i32");
-    assert!(!c_source.contains("T wrap"));
-    assert!(!c_source.contains("T unwrap"));
-}
+#[path = "multifile_generated_c/enum_dependencies.rs"]
+mod enum_dependencies;
+#[path = "multifile_generated_c/method_worklist_dependencies.rs"]
+mod method_worklist_dependencies;
